new decision2 package renamed to decision
