+++ internal/handler/decision/module.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package decision2
+package decision
 
 import (
 	"context"
