+++ internal/handler/decision/request_context_test.go
@@ -1,4 +1,4 @@
-package decision2
+package decision
 
 import (
 	"context"
